requirements.txt
openai==1.6.1
pandas==2.1.3
openpyxl==3.1.2
pillow==10.1.0
pdf2image==1.16.3
python-dotenv==1.0.0
pydantic==2.5.2
rich==13.7.0
typer==0.9.0
